git_ui: Show missing right border on selected items (#44747)
For folders and files basically any selected item in the git panel we
draw a border around it. The issue is that the right side of this border
wasn't ever visible.

In the project_panel.rs file I've saw that the decision was to make the
right side border 2 pixels. And this panel doesn't have this issue, no
matter which side of the dock is selected. So it was a very easy `look
at how we did x do y`.


Before: 

![image](https://github.com/user-attachments/assets/8ce32728-8ad6-487c-80f5-1c46d9756f4a)
After: 

![image](https://github.com/user-attachments/assets/998899b4-af98-4cc2-9435-4df6c98c1a50)

I don't think it warrants a release note. 

Release Notes:

- N/A

---------